Move stuff to util.go
